新功能：增加实验室新增或者修改时可修改或新增开放时间

+++ src/main/java/com/cuit9622/olms/entity/LabSchedule.java
@@ -23,7 +23,7 @@ public class LabSchedule implements Serializable {
      */
     @TableId
     @ApiModelProperty("实验室id")
-    private Integer labId;
+    private Long labId;
 
     /**
      * 1代表周日，其余以此类推
